Number of fixes and improvements
1. Removed support of Ubuntu 12.04
2. Removed python packages that are installed through virtualenv
3. Removed LLVM 3.3 (now installed through Impala toolchain)
4. Fixed LZO build

+++ cookbooks/impala/recipes/packages.rb
@@ -19,7 +19,7 @@ when "debian"
          "automake", "libtool", "flex", "bison", "cmake", "pkg-config", "git",
          "libssl-dev", "subversion", "libevent1-dev", "libsasl2-dev", "libldap2-dev",
          "liblzo2-dev", "lzop", "maven", "libboost-all-dev", "ccache", "python-pycurl",
-	 "libgnutls-dev"]
+	 "libgnutls-dev", "libgnutls-dev", "python-pycurl", "wget"]
     packages.each do |pkg|
       package pkg
     end
@@ -28,51 +28,24 @@ when "debian"
       user 'root'
       code <<-EOH
       cd /usr/lib/x86_64-linux-gnu
-      if ! test -f libboost_filesystem-mt.so; then
+      if ! test -h libboost_filesystem-mt.so && test -f libboost_filesystem.so; then
         sudo ln -s libboost_filesystem.so libboost_filesystem-mt.so
       fi
-      if ! test -f libboost_filesystem-mt.a; then
+      if ! test -h libboost_filesystem-mt.a && test -f libboost_filesystem.a; then
         sudo ln -s libboost_filesystem.a libboost_filesystem-mt.a
       fi
-      if ! test -f libbost_system-mt.so; then
+      if ! test -h libbost_system-mt.so && test -f libboost_system.so; then
         sudo ln -s libboost_system.so libboost_system-mt.so
       fi
-      if ! test -f libboost_system-mt.a; then
+      if ! test -h libboost_system-mt.a && test -f libboost_system.a; then
         sudo ln -s libboost_system.a libboost_system-mt.a
       fi
-      if ! test -f libbost_regex-mt.so; then
+      if ! test -h libbost_regex-mt.so && test -f libboost_regex.so; then
         sudo ln -s libboost_regex.so libboost_regex-mt.so
       fi
-      if ! test -f libboost_regex-mt.a; then
+      if ! test -h libboost_regex-mt.a && test -f libboost_regex.a; then
         sudo ln -s libboost_regex.a libboost_regex-mt.a
       fi
-      EOH
-    end
-  when node['platform_version'] == '12.04'
-    packages = ["build-essential", "ant", "libboost-dev", "libboost-thread-dev",
-                "libboost-test-dev", "libboost-program-options-dev", "libboost-regex-dev",
-                "libboost-system-dev", "libboost-filesystem-dev", "zlib1g-dev",
-                "libbz2-dev", "python-dev", "automake", "libtool", "flex", "bison",
-                "cmake", "pkg-config", "git", "libssl-dev", "subversion", "libevent1-dev",
-                "libsasl2-dev", "libldap2-dev", "libdb4.8-dev", "ccache"]
-    packages.each do |pkg|
-      package pkg
-    end
-
-    # Install Maven3 from an external PPA
-    apt_repository "maven3" do
-      uri 'ppa:natecarlson/maven3'
-      components ['main']
-      keyserver 'keyserver.ubuntu.com'
-      action :add
-    end
-
-    package "maven3"
-
-    bash "setup_maven" do
-      user 'root'
-      code <<-EOH
-      ln -s /usr/bin/mvn3 /usr/bin/mvn
       EOH
     end
   end
@@ -81,9 +54,7 @@ end
 # Python packages
 include_recipe "python::pip"
 
-python_pkgs = ["allpairs", "pytest", "pytest-xdist", "paramiko", "texttable",
-	       "prettytable", "sqlparse", "pywebhdfs", "gitpython", "jenkinsapi",
-	       "python-jenkins"]
+python_pkgs = ["python-jenkins"]
 
 python_pkgs.each do |pkg|
   python_pip pkg do
@@ -91,18 +62,14 @@ python_pkgs.each do |pkg|
   end
 end
 
-python_pip "psutil" do
-  version "0.7.1"
-end
-
 # Setup LZO
 bash 'setup_lzo' do
   user 'root'
   code <<-EOH
-  if ! test -f /usr/lib/liblzo2.a; then
+  if ! test -h /usr/lib/liblzo2.a && test -f /usr/lib/x86_64-linux-gnu/liblzo2.a; then
     ln -s /usr/lib/x86_64-linux-gnu/liblzo2.a /usr/lib/liblzo2.a
   fi
-  if ! test -f /usr/lib/liblzo2.so; then
+  if ! test -h /usr/lib/liblzo2.so && test -f /usr/lib/x86_64-linux-gnu/liblzo2.so; then
     ln -s /usr/lib/x86_64-linux-gnu/liblzo2.so /usr/lib/liblzo2.so
   fi
   EOH
